+++ LibTrixi.jl/.gitignore
@@ -1,2 +1,3 @@
 Manifest.toml
 run
+lib/init.o
